Move `get_last_assistant_message`
